backend/requirements.txt
fastapi
uvicorn
proxmoxer
websockets
httpx
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
